// shared-types/enums/mobile-surface-key.enum.ts
export enum MobileSurfaceKey {
  HOME = 'HOME',
  PRODUCT_CARD = 'PRODUCT_CARD',
  AUCTIONS_LIST = 'AUCTIONS_LIST',
  LISTING_CREATE = 'LISTING_CREATE',
  BUY_NOW = 'BUY_NOW',
  PRODUCT_DETAIL = 'PRODUCT_DETAIL',
  PROFILE = 'PROFILE',
  SETTINGS = 'SETTINGS',
  HOME_QUICK_TAB_BAR = 'HOME_QUICK_TAB_BAR',
}